Removed the base manager class
