+++ include/metall/metall.hpp
@@ -6,16 +6,16 @@
 #ifndef METALL_METALL_HPP
 #define METALL_METALL_HPP
 
-#include <metall/v0/manager_v0.hpp>
+#include <metall/v0/basic_manager.hpp>
 
 namespace metall {
 
 /// \brief Basic Metall manager type
 template <typename chunk_no_type, std::size_t chunk_size, typename kernel_allocator_type>
-using basic_manager = v0::manager_v0<chunk_no_type, chunk_size, kernel_allocator_type>;
+using basic_manager = v0::basic_manager<chunk_no_type, chunk_size, kernel_allocator_type>;
 
 /// \brief Metall manager type
-using manager = v0::manager_v0<>;
+using manager = v0::basic_manager<>;
 
 } // namespace metall
