+++ test/v0/manager_test.cpp
@@ -19,7 +19,7 @@ using namespace metall::detail;
 
 using chunk_no_type = uint32_t;
 static constexpr std::size_t k_chunk_size = 1 << 21;
-using manager_type = metall::v0::manager_v0<chunk_no_type, k_chunk_size>;
+using manager_type = metall::v0::basic_manager<chunk_no_type, k_chunk_size>;
 template <typename T>
 using allocator_type = typename manager_type::allocator_type<T>;
